sev_statement/requirements.txt
affine==2.2.2
amqp==2.5.1
appdirs==1.4.3
attrs==19.1.0
autopep8==1.4.4
Babel==2.7.0
billiard==3.6.1.0
cached-property==1.5.1
cairocffi==1.1.0
CairoSVG==2.4.0
celery==4.3.0
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
Click==7.0
click-plugins==1.1.1
cligj==0.5.0
coreapi==2.3.3
coreschema==0.0.4
cssselect2==0.2.1
cycler==0.10.0
defusedxml==0.6.0
dicttoxml==1.7.4
Django==2.1.15
django-celery-results==1.1.2
django-cors-headers==3.0.1
django-debug-toolbar==2.0
django-filter==2.1.0
django-model-utils==4.0.0
django-mssql==1.8
##django-notifications-hq==1.5.0
##django-pyodbc==1.1.3
##django-pyodbc-azure==2.1.0.0
django-rest-auth==0.9.5
django-rest-registration==0.5.0
django-rest-swagger==2.2.0
django-silk==3.0.4
django-simple-history==2.7.3
django-wkhtmltopdf==3.2.0
djangorestframework==3.9.4
djangorestframework-datatables==0.5.0
docopt==0.6.2
docxtpl==0.6.3
drf-flex-fields==0.6.1
drf-yasg==1.17.0
EasyProcess==0.2.7
ffmpeg-python==0.2.0
flower==0.9.3
future==0.18.2
gprof2dot==2019.11.30
gunicorn==19.9.0
html5lib==1.0.1
idna==2.8
image==1.5.27
importlib-metadata==0.23
inflection==0.3.1
isodate==0.6.0
itypes==1.1.0
Jinja2==2.10.1
jsonfield==2.0.2
kiwisolver==1.1.0
kombu==4.6.5
lxml==4.4.1
Markdown==3.1
MarkupSafe==1.1.1
matplotlib==3.1.1
money==1.3.0
more-itertools==7.2.0
num2words==0.5.10
numexpr==2.6.9
numpy==1.16.4
openapi-codec==1.3.2
opencv-contrib-python==4.1.1.26
packaging==19.2
pathlib==1.0.1
pbr==5.4.4
pdf2image
pdfkit==0.6.1
Pillow==6.0.0
preview-generator==0.13
psycopg2==2.8.2
py==1.4.22
pycodestyle==2.5.0
pycparser==2.19
pyexifinfo==0.4.0
Pygments==2.5.2
pyparsing==2.4.0
PyPDF2==1.26.0
Pyphen==0.9.5
pytest==2.6.0
python-dateutil==2.8.0
python-docx==0.8.7
python-magic==0.4.15
pytz==2019.1
PyVirtualDisplay==0.2.4
PyYAML==5.1.2
redis==3.3.11
requests==2.22.0
requests-toolbelt==0.9.1
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.2.0
simplejson==3.16.0
six==1.12.0
snuggs==1.4.6
sqlparse==0.3.0
stevedore==1.31.0
tinycss2==1.0.2
tornado==5.1.1
unoconv==0.9.0
uritemplate==3.0.0
urllib3==1.25.7
uWSGI==2.0.18
vine==1.3.0
virtualenv==16.7.9
virtualenv-clone==0.5.3
vtk==8.1.2
Wand==0.5.9
WeasyPrint==47
webencodings==0.5.1
xvfbwrapper==0.2.9
zeep==3.4.0
zipp==0.6.0
